+++ src/odoo_configurator/apps/import_configurator.py
@@ -96,7 +96,7 @@ class ImportConfigurator(base.OdooModule):
 
             xmlid = self.get_xmlid(model, record.get('id'))
             if not xmlid:
-                xmlid = self.compute_xml_id(record)
+                xmlid = self.compute_xml_id(record, model)
             res += '\n%s%s:' % (" " * 4 * 2, rec_name)
             res += '\n%s%s: %s' % (" " * 4 * 3, 'model', model)
             res += '\n%s%s: %s' % (" " * 4 * 3, 'force_id', xmlid)
@@ -144,6 +144,7 @@ class ImportConfigurator(base.OdooModule):
         field_name = field.get('name')
         model = field.get('model')
         field_type = field.get('ttype')
+        name = ''
         if field_type in ['one2many']:
             name = ''
         elif field_type in ['binary']:
@@ -159,15 +160,13 @@ class ImportConfigurator(base.OdooModule):
             files.append((file_name, binary_data))
             name = "\n%s%s: %s" % (" " * 4 * 4, field_name, 'get_image_local("%s")' % file_name)
         elif field_type == 'many2many':
-            # Todo : import many2many values
-            # rec_values = ''
-            # xmlid_list = sort_xml_ids([val.get_xml_id()[val.id] for val in record[field_name]])
-            # for xmlid in xmlid_list:
-            #     if xmlid:
-            #         rec_values += '\n%s- %s' % (" " * (2 + 4 * 4), xmlid)
-            # if rec_values:
-            #     name = '\n%s%s/id: %s' % (" " * 4 * 4, field_name, rec_values)
-            name = ''
+            rec_values = ''
+            xmlid_list = [self.get_xmlid(field['relation'], val) for val in record[field_name]]
+            for xmlid in xmlid_list:
+                if xmlid:
+                    rec_values += '\n%s- %s' % (" " * (2 + 4 * 4), xmlid)
+            if rec_values:
+                name = '\n%s%s/id: %s' % (" " * 4 * 4, field_name, rec_values)
 
         elif field_type in ['many2one']:
             if record[field_name]:
@@ -233,13 +232,13 @@ class ImportConfigurator(base.OdooModule):
         self.rec_to_xmlid_cache[(model, res_id)] = xmlid
         return xmlid
 
-    def compute_xml_id(self, record, retry=0):
-        xml_id = 'external_config.%s_%s' % (record['model'].replace('.', '_'), str(record['id']+retry).zfill(5))
+    def compute_xml_id(self, record, model, retry=0):
+        xml_id = 'external_config.%s_%s' % (model.replace('.', '_'), str(record['id']+retry).zfill(5))
         if not self._connection.get_id_from_xml_id(xml_id, no_raise=True):
             return xml_id
         else:
             retry += 1
-            return self.compute_xml_id(record, retry)
+            return self.compute_xml_id(record, model, retry)
 
     def apply(self):
         super(ImportConfigurator, self).apply()